implement rtk on auth components and on homepage
