+++ client/src/components/Carousal.jsx
@@ -4,42 +4,48 @@ import clock from "../../public/imgs/slider/language2.png";
 import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-
-import {
-  getAllCategory,
-  getCourseByCategory,
-} from "../services/operations/addCourses";
+import { setCourses } from "../store/reducers/coursesReducer";
+import { setCategories } from "../store/reducers/adminCategoryReducer";
 import CourseSkeleton from "./skeletons/CourseSkeleton";
+import {
+  useGetAllCategoryQuery,
+  useGetCategoryCourseQuery,
+} from "../services/course.api";
 
 export default function Carousal() {
   const dispatch = useDispatch();
 
-  // otherwise it is called again and again
-  useEffect(() => {
-    dispatch(getAllCategory());
-  }, []);
-
   const categories = useSelector((state) => state.categories.categories);
   const coursesAll = useSelector((state) => state.courses.courses.courses);
 
-  const [activeTab, setActiveTab] = useState(null); // Start with null or a default value
+  const [activeTab, setActiveTab] = useState(null);
+
+  const { data: allCategories } = useGetAllCategoryQuery();
+  const { data: categoryCourse } = useGetCategoryCourseQuery(activeTab, {
+    skip: !activeTab,
+  });
+
+  useEffect(() => {
+    if (allCategories?.success) {
+      dispatch(setCategories(allCategories.categories || []));
+    }
+  }, [allCategories, dispatch]);
 
   useEffect(() => {
     if (categories.length > 0 && !activeTab) {
-      setActiveTab(categories[0]._id); // Set the default value when categories are fetched
+      setActiveTab(categories[0]._id);
     }
-  }, [categories]); // Run this effect whenever categories change
+  }, [categories]);
 
   useEffect(() => {
-    if (activeTab) {
-      dispatch(getCourseByCategory(activeTab));
+    if (activeTab && categoryCourse?.success) {
+      dispatch(setCourses(categoryCourse.courses));
     }
-  }, [activeTab]);
+  }, [activeTab, categoryCourse, dispatch]);
 
   return (
     <div className="p-8 mt-4 w-full mx-auto">
-      {/* Title and Subtitle */}
-      <div className=" mb-6 items-center">
+      <div className="mb-6 items-center">
         <h2 className="font-sans font-semibold text-2xl text-[var(--color-primary)]">
           All the skills you need in one place
         </h2>
@@ -50,25 +56,22 @@ export default function Carousal() {
       </div>
 
       {/* Tab Menu */}
-      <div className="flex w-auto gap-4 space-x-1 overflow-x-auto carousel  scroll-snap-x scroll-smooth ">
-        {categories
-          .filter((_, i) => i < 4)
-          .map((tab) => (
-            <button
-              key={tab._id}
-              onClick={() => {
-                setActiveTab(tab._id);
-              }}
-              className={`flex justify-center items-center w-fit px-0 py-2 border-b-2 cursor-pointer ${
-                activeTab === tab._id
-                  ? " border-[var(--color-primary)] text-[var(--color-primary)]"
-                  : " border-transparent text-gray-600 carousel-item"
-              }`}
-            >
-              {tab.categoryName}
-            </button>
-          ))}
+      <div className="flex w-auto gap-4 space-x-1 overflow-x-auto carousel scroll-snap-x scroll-smooth">
+        {categories.slice(0, 4).map((tab) => (
+          <button
+            key={tab._id}
+            onClick={() => setActiveTab(tab._id)}
+            className={`flex justify-center items-center w-fit px-0 py-2 border-b-2 cursor-pointer ${
+              activeTab === tab._id
+                ? " border-[var(--color-primary)] text-[var(--color-primary)]"
+                : " border-transparent text-gray-600 carousel-item"
+            }`}
+          >
+            {tab.categoryName}
+          </button>
+        ))}
       </div>
+
       <div className="relative w-[80%]">
         <hr className="border-gray-200" />
         <Link
@@ -80,11 +83,11 @@ export default function Carousal() {
       </div>
 
       {/* Scrollable Course Cards */}
-      {coursesAll ? (
+      {Array.isArray(coursesAll) && coursesAll.length > 0 ? (
         <div className="flex items-center flex-wrap md:flex-nowrap gap-5 p-4">
-          {coursesAll?.map((course, index) => (
+          {coursesAll.map((course) => (
             <Link
-              key={index}
+              key={course._id}
               to={`/course/${course._id}`}
               className="bg-white flex flex-col gap-2 w-[296px] pb-3 rounded-lg shadow-md"
             >
@@ -104,7 +107,7 @@ export default function Carousal() {
                     <img
                       src={user}
                       alt="userLive"
-                      className="w-[25px] h-[25px] text-blue-950"
+                      className="w-[25px] h-[25px]"
                     />
                     <p className="text-gray-600">
                       {course.courseMode.charAt(0).toUpperCase() +
