+++ client/src/modals/UpdatePasswordModal.jsx
@@ -1,21 +1,21 @@
-import { useEffect, useRef, useState } from "react";
 import { RxCross2 } from "react-icons/rx";
-import { useDispatch } from "react-redux";
+
+import { toast } from "react-hot-toast";
+import { useEffect, useRef, useState } from "react";
 import ButtonLoading from "../components/Button/ButtonLoading";
-import { updatePassword } from "../services/operations/authApi";
+import { useUpdatePasswordMutation } from "../services/auth.api";
 
 function UpdatePasswordModal({
   email,
   updatePasswordModal,
   setUpdatePasswordModal,
-  setLoginModal,
 }) {
   const modalRef = useRef(null);
   const otpInputRefs = useRef([]);
-  const dispatch = useDispatch();
+
+  const [updatePassword, { isLoading, error }] = useUpdatePasswordMutation();
 
   const [isOtpComplete, setIsOtpComplete] = useState(false);
-  const [loading, setLoading] = useState(false);
   const [userData, setUserData] = useState({
     email: email,
     otp: "",
@@ -83,13 +83,18 @@ function UpdatePasswordModal({
     setUserData((prev) => ({ ...prev, [e.target.name]: e.target.value }));
   };
 
-  const otpSubmitHandler = (e) => {
+  const otpSubmitHandler = async (e) => {
     e.preventDefault();
     const otpValues = otpInputRefs.current.map((ref) => ref.value).join("");
     const userRegisterData = { ...userData, otp: otpValues };
-    dispatch(
-      updatePassword(userRegisterData, setUpdatePasswordModal, setLoginModal)
-    );
+
+    try {
+      await updatePassword(userRegisterData);
+      toast.success("Password Updated Successfully");
+      setUpdatePasswordModal(false);
+    } catch (err) {
+      toast.error(error);
+    }
   };
 
   return (
@@ -146,13 +151,13 @@ function UpdatePasswordModal({
 
         <button
           className={`w-full p-2 rounded-lg transition ${
-            loading
+            isLoading
               ? "bg-gray-500 cursor-not-allowed"
               : "bg-blue-600 hover:bg-blue-700"
           }`}
-          disabled={!isOtpComplete || loading}
+          disabled={!isOtpComplete || isLoading}
         >
-          {loading ? <ButtonLoading /> : "Submit"}
+          {isLoading ? <ButtonLoading /> : "Submit"}
         </button>
       </form>
     </div>
